Add: Multi lines command, memberreg, start message | Changed: reg databases call names

+++ system/commands/pack/profilereqaccept/main.js
@@ -4,30 +4,37 @@ const   Dictionary  = require(global.PROJECTDIR+'botdictionary.json');
 // const   Mongo       = require('mongodb').MongoClient; // Удалить на релизе
 // const MongoClient   = new Mongo(MongoCFG.url, {useUnifiedTopology: true}); // Удалить на релизе
 
+
 module.exports =
 function(arg, _, aliascommand){
-    let database = arg.MongoClient.db(MongoCFG.regdb); // добавить arg. к MongoClient на релизе
-    let collectionlist = database.collection(MongoCFG.collregreq);
+    let database = arg.MongoClient.db(MongoCFG.dbreg); // добавить arg. к MongoClient на релизе
+    let collectionlist = database.collection(MongoCFG.collregprofilereq);
     
     // Если аргумент hash отсутствует
     if (!aliascommand.commandargs.length) return arg.msg.reply(Dictionary.errors.wronrarg);
 
     collectionlist.findOne({hash: aliascommand.commandargs[0]}, (err, document) =>{
         if (err){
-            let errmsg = Dictionary.errors.mongodberror.replace("#00", "#08"); 
+            let errmsg = Dictionary.errors.mongodberror.replace("#00", "#20"); 
             return arg.msg.reply(errmsg);
         }
         // Если null - неправильный hash 
         if (document != null){
+            let role = arg.msg.guild.roles.cache.find((value) => {
+                return value.name == document.role;
+            });
+
+            if (role === undefined) return arg.msg.reply(Dictionary.errors.reqaccrolenotfound);
+
             collectionlist.deleteOne(document, (_err, _res) => {
                 
                 if (err){
-                    let errmsg = Dictionary.errors.mongodberror.replace("#00", "#09"); 
+                    let errmsg = Dictionary.errors.mongodberror.replace("#00", "#21"); 
                     return arg.msg.reply(errmsg);
                 }
                 
                 if (_res.result.ok != 1){
-                    let errmsg = Dictionary.errors.mongodberror.replace("#00", "#10"); 
+                    let errmsg = Dictionary.errors.mongodberror.replace("#00", "#23"); 
                     return arg.msg.reply(errmsg);
                 }
 
@@ -37,26 +44,23 @@ function(arg, _, aliascommand){
                 
                 // Если найден - находится в канале
                 if (user){
-                    let replyDMmsg = Dictionary.DM.regdenymsg.replace("#profile", document.profilename);
-                    if (aliascommand.commandargs[1] !== undefined) {
-                        let reason = aliascommand.commandargs.slice(1); 
-                        reason = reason.join(" ");
-                        replyDMmsg += "\nПричина: " + reason;
-                    }
+                    let replyDMmsg = Dictionary.DM.regaccmsg.replace("#profile", document.profilename);
+                        user.roles.add(role);
                     // Может быть ошибка, по причине закрытых сообщений
                     try{
                         user.send(replyDMmsg);
-                        let replymsg = Dictionary.reply.reqdeny.replace("#hash", aliascommand.commandargs[0]);
+
+                        let replymsg = Dictionary.reply.reqacc.replace("#hash", aliascommand.commandargs[0]);
                         replymsg += "\n" + Dictionary.additional.reqnotified;
                         arg.msg.reply(replymsg);
                     }catch(e){
-                        let replymsg = Dictionary.reply.reqdeny.replace("#hash", aliascommand.commandargs[0]);
+                        let replymsg = Dictionary.reply.reqacc.replace("#hash", aliascommand.commandargs[0]);
                         replymsg += "\n" + Dictionary.additional.reqcloseddm;
                         arg.msg.reply(replymsg);
                     }
                 }
                 else{
-                    let replymsg = Dictionary.reply.reqdeny.replace("#hash", aliascommand.commandargs[0]);
+                    let replymsg = Dictionary.reply.reqacc.replace("#hash", aliascommand.commandargs[0]);
                     replymsg += "\n" + Dictionary.additional.reqnotmember;
                     arg.msg.reply(replymsg);
                 }
@@ -64,5 +68,4 @@ function(arg, _, aliascommand){
         }
         else return arg.msg.reply(Dictionary.errors.reqhashnotfound);
     });
-    
 }
